pyproject.toml
[project]
name = "mHC"
version = "0.1.0"
description = "Implementing Manifold Hyper Connections for Image Classigication"
readme = "README.md"
requires-python = "==3.9.21"
dependencies = [
    "black>=25.11.0",
    "ipykernel>=6.30.1",
    "isort>=6.1.0",
    "jupyter>=1.1.1",
    "kagglehub>=0.3.12",
    "lab>=8.6",
    "matplotlib>=3.9.4",
    "pandas>=2.3.1",
    "pytest>=8.4.2",
    "pytorch-warmup>=0.2.0",
    "setuptools>=82.0.0",
    "tensorboard>=2.20.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.23.0",
]
